requirements.txt
streamlit>=1.20.0
pandas>=1.3.0
numpy
plotly
openai
pytz
mysql-connector-python
streamlit_extras==0.5.0
streamlit-cookies-manager
streamlit-cookies-controller
tvdatafeed @ git+https://github.com/rongardF/tvdatafeed.git